Rename io modules to io_ and fix some version conflicts (#78)
* Rename io modules to io_ and fix some version conflicts

Numpy 1.16.* is not compatible with sparsity 0.20.* thus we need to fix
the setup.py. When using Scipy<1.0.0 empty column access does not work,
thus the dependency had to be adjusted here as well.
This also renames the io_ modules to avoid issues with pythons
internal module.

* Fix incompatibility with numpy>=1.16.0, potential security issue.

Due to a security issue (CVE-2019-6446) numpy changed the default value
 of allow_pickle in np.load to True, this led to error when reading
 sparse frames from npz archives. This commit fixes it by allowing
 pickled objects, thus reading sparse frames from unkown sources is
 still a security risk.
